// src/redux/common/types.ts
export enum ActionType {
  INIT_APP = 'COMMON/INIT_APP',
  INIT_APP_SUCCESS = 'COMMON/INIT_APP_SUCCESS',
  ERROR = 'COMMON/ERROR',
  INFO = 'COMMON/INFO',
  SUCCESS = 'COMMON/SUCCESS',
  DISMISS_MESSAGE = 'COMMON/DISMISS_MESSAGE',
  PREVIEW_IMAGES = 'COMMON/PREVIEW_IMAGES',
  PREVIEW_END = 'COMMON/PREVIEW_END',
  SET_TOKEN = 'COMMON/SET_TOKEN',
}